Finished basic functionality of user controller
