+++ index.js
@@ -1,7 +1,7 @@
 const express = require('express')
 const morgan = require('morgan')
 const db = require('./src/models')
-const empleado_routes = require("./src/routes/empleado-routes");
+const user_routes = require("./src/routes/user-routes");
 const home_routes = require("./src/routes/home-routes");
 
 const app = express()
@@ -18,7 +18,7 @@ db.sequelize.sync()
 app.use(morgan('dev'))
 app.use(express.json());
 
-app.use("/user", empleado_routes);
+app.use("/user", user_routes);
 app.use(home_routes);
 
 app.all('*', (req, res) => {
