--- src/routes/empleado-routes.js
@@ -1,10 +1,0 @@
-// Router para empleado:
-let express = require("express");
-let empleado_controller = require("../controllers/empleado-controller");
-
-let router = express.Router();
-
-router.get("/", empleado_controller.empleado_index);
-router.post("/", empleado_controller.empleado_crear);
-
-module.exports = router;
